requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   pydantic
anyio==4.5.2
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   starlette
asgiref==3.8.1
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   opentelemetry-instrumentation-asgi
click==8.1.8
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   uvicorn
deprecated==1.2.18
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
exceptiongroup==1.3.0
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   anyio
fastapi==0.115.12
    # via realtime-credit-card-fraud-detection (pyproject.toml)
h11==0.16.0
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   uvicorn
idna==3.10
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   anyio
importlib-metadata==8.5.0
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   opentelemetry-api
joblib==1.4.2
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   scikit-learn
numpy==1.24.4
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   pandas
    #   scikit-learn
    #   scipy
opentelemetry-api==1.33.1
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   opentelemetry-exporter-prometheus
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-prometheus==0.54b1
    # via realtime-credit-card-fraud-detection (pyproject.toml)
opentelemetry-instrumentation==0.54b1
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi==0.54b1
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.54b1
    # via realtime-credit-card-fraud-detection (pyproject.toml)
opentelemetry-sdk==1.33.1
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   opentelemetry-exporter-prometheus
opentelemetry-semantic-conventions==0.54b1
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
opentelemetry-util-http==0.54b1
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
packaging==25.0
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   opentelemetry-instrumentation
pandas==2.0.3
    # via realtime-credit-card-fraud-detection (pyproject.toml)
prometheus-client==0.21.1
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   opentelemetry-exporter-prometheus
pydantic==2.10.6
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   fastapi
pydantic-core==2.27.2
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   pydantic
python-dateutil==2.9.0.post0
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   pandas
pytz==2025.2
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   pandas
scikit-learn==1.3.2
    # via realtime-credit-card-fraud-detection (pyproject.toml)
scipy==1.10.1
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   scikit-learn
six==1.17.0
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   python-dateutil
sniffio==1.3.1
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   anyio
starlette==0.44.0
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   fastapi
structlog==25.4.0
    # via realtime-credit-card-fraud-detection (pyproject.toml)
threadpoolctl==3.5.0
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   scikit-learn
typing-extensions==4.13.2
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   annotated-types
    #   anyio
    #   asgiref
    #   exceptiongroup
    #   fastapi
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   starlette
    #   structlog
    #   uvicorn
tzdata==2025.2
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   pandas
uvicorn==0.33.0
    # via realtime-credit-card-fraud-detection (pyproject.toml)
wrapt==1.17.2
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   deprecated
    #   opentelemetry-instrumentation
zipp==3.20.2
    # via
    #   realtime-credit-card-fraud-detection (pyproject.toml)
    #   importlib-metadata
